// src/types/interfaces/FormTypes/ArrayOfFormsForm.ts
import { InputsArray } from 'types/interfaces/FormTypes/InputObject'
import GenericForm, { formTypes } from './GenericForm'
import { InputsFormValue } from 'types/interfaces/FormTypes/InputsForm'

export interface ArrayOfFormsForm
  extends GenericForm<ArrayOfFormsFormValue, unknown, boolean> {
  type: formTypes.INPUTS
  forms: InputsArray[]
  isAutoGenerated?: string
}

export type ArrayOfFormsFormValue = {
  principalForm: number
  formsValues: Record<string, InputsFormValue>[]
}

export type SubmitFormsType = {
  isValid: boolean
  value: ArrayOfFormsFormValue
}